webocr/requirements.txt
git+https://github.com/vertica/torndb-pymysql
git+https://github.com/vertica/schemaless
simplejson==3.8.1
pymysql==0.6.7
tornado==4.3
git+https://github.com/vertica/tornado_jinja2
celery==3.1.19
pytest==2.8.1
pytest-tornado==0.4.4
pytest-cov==2.2.0
futures==3.0.3
flake8==2.5.1
requests==2.9.1
python-magic==0.4.10

